python/requirements.txt
letta-client
python-socketio
requests